Introduced multiple files and directories, and the possibility to omit them, on start up.
Misc, error-checking. Removed defaults in directory and files.

a) -f (files flag) can now be used, to specify files instead of
directories, and multiple files is supported by using a ':'-delimeter.
A note on processing. When a file returns exit(>0) the processing
aborts.The field is optional.

b) -d (directory flag) now supports multiple directories, again using 
the ':'-delimiter on the command-line. When a file inside the directory
returns exit(>0) the processing of the directory stops. The field is
optional

- However, one of the above must be supplied, or you'll have one stupid
ssl-server.æ